Fix cost calculation for reverse order edges
