+++ tools/configs/uniform.ini
@@ -1,5 +1,5 @@
 [Input]
-input_dir=results_17_06
+input_dir=/home/pdominik/Tensor_experiments/results_val
 input_prefix=results
 
 [Output]
